+++ python/bindings/openravepy_global.cpp
@@ -108,7 +108,7 @@ public:
     }
 
     virtual string __repr__() {
-        return boost::str(boost::format("<AABB([%f,%f,%f],[%f,%f,%f])>")%ab.pos.x%ab.pos.y%ab.pos.z%ab.extents.x%ab.extents.y%ab.extents.z);
+        return boost::str(boost::format("AABB([%f,%f,%f],[%f,%f,%f])")%ab.pos.x%ab.pos.y%ab.pos.z%ab.extents.x%ab.extents.y%ab.extents.z);
     }
     virtual string __str__() {
         return boost::str(boost::format("<%f %f %f %f %f %f>")%ab.pos.x%ab.pos.y%ab.pos.z%ab.extents.x%ab.extents.y%ab.extents.z);
@@ -365,7 +365,7 @@ public:
 
     string __repr__() {
         std::stringstream ss;
-        ss << _spec;
+        ss << "ConfigurationSpecification(\"\"\"" << _spec << "\"\"\")";
         return ss.str();
     }
     string __str__() {
@@ -567,7 +567,7 @@ public:
         std::stringstream ss;
         ss << std::setprecision(std::numeric_limits<dReal>::digits10+1);     /// have to do this or otherwise precision gets lost
         ss << _param;
-        return boost::str(boost::format("<IkParameterization('%s')>")%ss.str());
+        return boost::str(boost::format("IkParameterization('%s')")%ss.str());
     }
     string __str__() {
         std::stringstream ss;